requirements.txt
# Core Framework
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-wtf==1.2.1
flask-migrate==4.0.5
flask-cors==4.0.0
flask-mail==0.9.1
flask-socketio==5.3.6

# Database
psycopg2-binary==2.9.9

# Data Science & ML
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
xgboost==2.0.3
lightgbm==4.1.0
prophet==1.1.5
statsmodels==0.14.1

# Optimization
ortools==9.8.3296

# Visualization
folium==0.15.1
plotly==5.18.0
seaborn==0.13.0
matplotlib==3.8.2

# API & Utilities
requests==2.31.0
python-dotenv==1.0.0
wtforms==3.1.1
email-validator==2.1.0

# Excel/CSV Export
openpyxl==3.1.2
reportlab==4.0.7

# Task Queue (Optional)
celery==5.3.4
redis==5.0.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
faker==21.0.0

# Production Server
gunicorn==21.2.0
gevent==23.9.1
